scene and text refactorting
