+++ src/CMakeLists.txt
@@ -43,4 +43,6 @@ add_subdirectory(
 	${CMAKE_CURRENT_SOURCE_DIR}/input
 )
 
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/helpers)
+
