+++ src/app/CMakeLists.txt
@@ -2,8 +2,8 @@ add_library(app SHARED)
 
 target_include_directories(app PUBLIC include)
 
-target_sources(app PRIVATE app.cpp text_rendering.cpp)
+target_sources(app PRIVATE app.cpp)
 
-target_link_libraries(app PUBLIC render_lib render_vulkan_lib window text_render_lib input_sdl2_lib)
+target_link_libraries(app PUBLIC render_lib render_vulkan_lib window text_render_lib input_sdl2_lib helpers)
 
 target_compile_definitions(app PRIVATE EXPORT_TZ_LIB)
