+++ src/helpers/include/text_helper.hh
@@ -1,0 +1,30 @@
+//
+// Created by Martin Gruscher on 19.05.26.
+//
+
+#ifndef SWARMS_TEXT_HELPER_HH
+#define SWARMS_TEXT_HELPER_HH
+#include <render_helpers.hh>
+
+namespace tz
+{
+
+class TextHelper
+{
+public:
+  TextHelper(render::vulkan::Renderer &renderer, TextureAssetManager &textureAssetManager);
+  int createFont(const std::string& fileName, int size);
+  void renderText(Transform transform, const std::string& text, int fontId = -1);
+
+private:
+  int uiFont = -1;
+  uint32_t uiFontAtlasTextureIndex = 0;
+  render::vulkan::Renderer& renderer;
+  render::TextRenderer * textRenderer = nullptr;
+  TextureAssetManager& textureAssetManager;
+};
+}
+
+
+
+#endif //SWARMS_TEXT_HELPER_HH
